fix compile issue of CarbonSessionCatalog for spark 2.4
